cro_crc_package_name_fixed
