+++ script-generator/src/main/java/com/coolbitx/wallet/signing/main/AllScript.java
@@ -16,8 +16,8 @@ public class AllScript {
         XlmScript.listAll();
         XrpScript.listAll();
         AdaScript.listAll();
-        CroScript.listAll();
         CrcScript.listAll();
+        CroScript.listAll();
     }
 
 }
